fix(status): keep behavior model under bundle budget
